requirements.txt
asgiref==3.11.0
Django==5.2.8
djangorestframework==3.16.1
numpy==2.3.5
pandas==2.3.3
psycopg==3.3.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
